LORE-256 Separate Delete into its own Controller class
